+++ DependencyInjection/Compiler/LoggerAutowireCompilerPass.php
@@ -13,9 +13,11 @@ declare(strict_types=1);
 
 namespace Linkin\Bundle\MonologAutowireBundle\DependencyInjection\Compiler;
 
+use Linkin\Bundle\MonologAutowireBundle\Cache\LoggerClassCache;
 use Linkin\Bundle\MonologAutowireBundle\Handler\LoggerHandler;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
 use function str_replace;
 use function strpos;
 
@@ -35,6 +37,9 @@ class LoggerAutowireCompilerPass implements CompilerPassInterface
 
         $loggerChannels = [];
 
+        $loggerCache = new LoggerClassCache();
+        $loggerCache->clear();
+
         foreach ($container->getDefinitions() as $id => $definition) {
             if (strpos($id, 'monolog.logger.') !== 0) {
                 continue;
@@ -43,6 +48,10 @@ class LoggerAutowireCompilerPass implements CompilerPassInterface
             $channelName = str_replace('monolog.logger.', '', $id);
 
             $loggerChannels[$channelName] = $definition;
+
+            $loggerFullClassName = $loggerCache->generateClass($channelName);
+
+            $container->register($loggerFullClassName, $loggerFullClassName)->addArgument(new Reference($id));
         }
 
         $container->getDefinition(LoggerHandler::class)->replaceArgument(0, $loggerChannels);